+++ src/components/card/comments-counter/comments-counter.module.css
@@ -1,0 +1,20 @@
+.commentsContainer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    background-color: rgb(47, 63, 97);
+    border-radius: 2rem;
+    width: 4rem;
+    height: 2rem;
+    transition: background-color 0.3s;
+    gap: 0.5rem;
+}
+
+.commentsContainer p {
+    color: white;
+}
+
+.commentsContainer:hover {
+    background-color: rgb(58 75 111);
+}
