--- src/components/comments-counter/comments-counter.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-interface CommentsCounterProps {
-    counter: number
-}
-
-export const CommentsCounter = ({counter}: CommentsCounterProps) => {
-    return <p>{counter}</p>
-}
