+++ src/components/generic/button/button.tsx
@@ -3,13 +3,13 @@ import styles from "./button.module.css";
 
 interface ButtonProps {
     text?: string;
-    onClick?: (e: React.MouseEvent<HTMLElement>) => void;
-    disabled?: boolean
-    pathD: string;
+    onClick?: (e: React.MouseEvent<HTMLElement>) => void,
+    disabled?: boolean,
+    pathD?: string,
 }
 
 export const Button = (props: ButtonProps) => {
-    const {pathD, onClick, disabled} = props;
+    const {pathD, onClick, disabled, text} = props;
     return (
         <>
             <button
@@ -18,10 +18,13 @@ export const Button = (props: ButtonProps) => {
                 onClick={onClick}
                 disabled={disabled}
             >
-                <svg fill="currentColor" height="16" viewBox="0 0 20 20" width="16"
-                     xmlns="http://www.w3.org/2000/svg">
-                    <path d={pathD} />
-                </svg>
+                {pathD && (
+                    <svg fill="currentColor" height="16" viewBox="0 0 20 20" width="16"
+                         xmlns="http://www.w3.org/2000/svg">
+                        <path d={pathD}/>
+                    </svg>
+                )}
+                {text}
             </button>
         </>
     )
