+++ src/components/generic/image/image.tsx
@@ -1,0 +1,22 @@
+import React, {useState} from 'react';
+
+interface ImageUtilsProps {
+    img: string,
+    defaultImage: string
+}
+
+export const Image = (props: ImageUtilsProps) => {
+    const {img, defaultImage} = props
+    const [imageError, setImageError] = useState(false);
+
+    const handleImageError = () => {
+        setImageError(true);
+    };
+    return (
+        <img
+            src={imageError ? defaultImage : img}
+            alt="default"
+            onError={handleImageError}
+        />
+    );
+}
