+++ src/features/main/content/all-cards.module.css
@@ -1,6 +1,5 @@
 .allCards {
     display: flex;
-    /*row*/
     flex-direction: column;
     align-items: center;
 }
