+++ src/utils/date-utils.ts
@@ -1,8 +1,8 @@
 export const getTimeDifference = (targetDate: Date) => {
     const currentDate = new Date();
     const currentTime = currentDate.getTime();
-    const createdDateTime = targetDate.getTime();
-    const timeDiff = currentTime - createdDateTime;
+    const targetDateTime = targetDate.getTime();
+    const timeDiff = currentTime - targetDateTime;
     const minutesPassed = Math.floor(timeDiff / (1000 * 60));
     const hoursPassed = Math.floor(timeDiff / (1000 * 60 * 60));
     const daysPassed = Math.floor(timeDiff / (1000 * 60 * 60 * 24));
